Rudimentary pragma mechanism added and tested
The spice source file (*.cir) can now contain pragmas that
are defined by starting the line with "*Pragma".
The rest of the line should contain a name, a value range and a
number of steps. For example

*Pragma foobar 5.0 6.5 10

This line is omitted from the circuit description that is sent to
Josim. However all subsequent instances of the string "~foobar~"
will be substituted with the number "5.0". In subsequent iterations,
this number will be increase up to and including "6.0" in 10 steps.